Fix javascript not executing after content load.
When we set the inner html of an element the scirpt tags are added
but not evaluated. This caused the facebook plugin to stop working
when the new content loading mechanim was used. To solve this we
remove the original script nodes and append them again. This will
force them to be executed.